change onChange api
