+++ src/utils/mapper.ts
@@ -1,6 +1,6 @@
-import { Toast, ToastItem } from '../types';
+import { Toast, ToastItem, ToastItemStatus } from '../types';
 
-export function toToastItem(toast: Toast): ToastItem {
+export function toToastItem(toast: Toast, status: ToastItemStatus): ToastItem {
   return {
     content: toast.content,
     containerId: toast.props.containerId,
@@ -9,6 +9,7 @@ export function toToastItem(toast: Toast): ToastItem {
     type: toast.props.type,
     data: toast.props.data || {},
     isLoading: toast.props.isLoading,
-    icon: toast.props.icon
+    icon: toast.props.icon,
+    status
   };
 }
